WIP: reworking recipe page
Basically rewriting the recipe page. Need to hook it up to the store.

Created an EnhancedInput element which allows a user to switch between
the presentational component and the editing component. Need to
generalize it.

Removed some bad practices with setting the state to a prop in the
constructor. Just use the prop silly!

Hooked up the Redux chrome extension

Setup an addStepToRecipe action with an accompanying test.