Add redirection when logging in
When a user attempts to navigate to an unauthorized address, like
`/recipes`, they are redirected to the Login page. Now, when they
sucessfully login, they will be redirected back to where they were
attempting to navigate to, when the login using email/password.

Need to add this functionality for social login, which will be more
challenging.