Remove backend-based searching and search modal
Searching on the frontend is cheap and easy to do. This feature has been broken
for ages and I believe it is not as responsive as the frontend search feature.

We can revisit this later if needed, but I believe the number of recipes for a
team will remain small enough that frontend search will remain sufficient.