Restrict DB access in serialization (#285)
N+1 queries are more of an issue with list views, so I think it's acceptable to
allow queries inside create views where we only serialize one object.

We only raise exceptions in DEBUG mode. If a serializer makes a DB call in production we log an error.

Add toast for the RecipeItem schedule button action.