Move poetry to root (#381)
This is the first step in moving all our tools to root.

We want to keep tooling at the root so we can apply JS and Python tools
to the whole repo.

I'm surprised poetry works correctly from a sub directory. This means we
don't need to update our calling procedure in all our scripts.