Move format to root. Remove unused shebang search feature.
Moving format script to root. I've removed the unused feature that
searches for extensionless python files based on their shebang. It's a
neat feature, but the feature I added it for no longer exists, so I'm
removing it to improve clarity.

Add configuration to pyproject.toml so black ignores node_modules