add(cumin): better handling of unknown units
Previously we were largely ignoring unknown units which wasn't great as
the rendered shopping list had some gaps.

Ideally we'd have the ingredients be something like

enum Unit {
    TABLESPOON,
    // ...
    UNKNOWN(String)
}

But alas, we can't easily achieve that so instead we add a `unknown_unit`
field to the Quantity() which is nullable.

Also round decimals to prevent issues with combining resulting in:

`6.500000000000000000000000002` Tablespoons