add(api): ical endpoint for syncing scheduled recipes to calendar apps
Use icalendar https://tools.ietf.org/html/rfc5545 to sync scheduled
recipes to apps like Apple Calendar.

We limit the length of the ical file to the last year of scheduled
recipes to prevent an unbounded file size. This means the calendar app
will delete the scheduled recipes that are older than this cutoff which
is less than ideal, but not sure the way around it.

We use a random string instead of a UUID so we can have a larger
alphabet and also a greater number of characters to get more randomness.
Ideally we'd have each team member get their own icalendar link & secret
so that if they leave the team they don't keep access to the calendar.