ref(ui): ingredient component to use hooks
This is in preparation for adding positioning & drag and drop to
ingredients along with sections. The hooks api for React DND is much
nicer to work with.

I also changed the export from a default to a named export.

Also disable the tslint rule since `type` has inferred index signatures
making it support more features than interface.