add: ingredient sections (#580)
Add sections for ingredients, so for more complicated recipes we can
section off parts of the ingredients, e.g. _for the sauce_

We add a new model which is keeping the same float setup for position
because that's easier than setting up the string positioning system.

Also the position is somewhat "shared" across the ingredient model since
they exist in the same list, and are sorted together. This results in
some weirdness with handlers and sorting but it works out.