fix(ical): random unique ids on each fetch
The events used in the ical calendar each have unique ids, but they are
randomly generated on each request. This is probably what was causing
the glitchiness in the iOS calendar app (it would start scrolling back
up in history when it fetched the calendar).

So now we use the ScheduledRecipe id along with a prefix of the tablename
to help ensure uniqueness.