fix(ui): hover handling for calendar items
We were depending on the mouseEvent and mouseLeave events to fire to
keep track of the hover state for scheduled recipes.

This works okay when deleting a single item as when you drag your mouse
over to the item, the onMouseEnter handler fires.

However, when you have a couple items on the same day and delete the
first one, the next item will move up in the calendar day and the event
handlers won't fire. So the component will have its hover state set to
false when we are in fact hovering.

Instead of keeping track of hover state ourselves, we instead query the
DOM with `.matches()`

https://developer.mozilla.org/en-US/docs/Web/API/Element/matches