add(ui): linking directly to notes in recipes
On component mount we check if the hash is equal to the note id, if it
is we add the class to highlight the comment.

Initially tried using a `:target` pseudo selector approach but that
didn't end up working because we aren't SSR so we have a spinner and
the browser doesn't highlight the `:target` if it isn't initially in
the html it seems.