fix(ui): error w/ inflight request while naving to diff hostname on iOS
There's been this persistent issue on iOS since we added more network
requests to sync the recipe data.

When naving away there will be an infight network request due to some
onfocus handlers, this network request gets killed by the browser with
an error: NetworkError

This was surfaced in the UI as a 404.

Now we return early from the network request. Really we want to show an
error unless a previous network request has succeeded, but
this is fine for now.