fix(ui): form submission bug with ingredient (#696)
We recently reordered the buttons in most of the forms, which seems
innocuous but caused a bug where, when hitting Enter inside the update
ingredient form, the ingredient would be deleted instead of updated.

This is due to the Delete button being first in the html form and not
having a `type` set which defaulted it to being `type=submit`. So Enter
was causing that button to be clicked, deleting the ingredient.

rel: https://stackoverflow.com/a/4763911/3720597

Also delete some unused error props.