add(shopping list): naive sorting of ingredients
Sub string matching of a bunch of manually created strings.

I think a future version would use some stemming at least and some
similarity metrics.

Also we calculate this on the fly instead of storing it in the database.
In the future we may want to persist this property to the database.

For testing I dumped the database and manually updated the category
function into everything was categorized.

In the future we can update this CSV. I used the following query to
create the CSV:

```sql
SELECT DISTINCT
	i.name AS name
FROM
	core_ingredient i
	JOIN core_recipe AS r ON r.id = i.recipe_id
WHERE
	i.deleted_at IS NULL
	AND r.deleted_at IS NULL
	AND r.archived_at IS NULL
	AND r.object_id = 2

order by i.name;
```

Also to fix some spelling issues on prod I used this query to find
the recipes:

```sql
SELECT
	i.name,
	i.quantity,
	r.name,
	r.archived_at,
	r.deleted_at
FROM
	core_ingredient i
	JOIN core_recipe AS r ON r.id = i.recipe_id
WHERE
	i.name = 'kosher sealt';
```

To find `"unknown"` ingredients, I think the following should work in
the future:

```sql
SELECT
	json_data.key AS key,
	json_data.value AS value,
	json_data.value ->> 'category'
FROM
	core_shoppinglist,
	jsonb_each((core_shoppinglist.ingredients::jsonb ->> 0)::jsonb) AS json_data
```