fix(ui): image upload flashing (#759)
During the upload process, we'd swap the local id of the uploaded image
with the server id. This should be fine, except we use the id for a
`key` in an array, so that causes the `<img` tag to unmount, resulting
in a flash as the new image loads in.

Normally if you change the `src` of an image, the browser will wait for
the new one to load before swapping it out, but since we were using a
`key`, the `<img` tag was being deleted and added back to the DOM.

Before realizing `key` was the issue, was going to go the route of
preloading the image with:

```ts
const preload = (src: string) => {
  const img = new Image()
  img.src = src
  return img.decode()
}
```

which worked, but fixing the key issue seems better although not sure.

PS: also added error handling for when the upload API fails.