fix(api): primary image backfill (#971)
we were skipping images that weren't attached to notes, aka ones that we scrape for the initial primary image

also I think rendering a background-image with an empty string url results in some glitchy looking UI but not 100% sure. So now we don't include the css property when we don't have a backgroundImage url.