ui: fix background blur for image
I'm not sure why, but backdrop-filter seems to get prefix on dev but not in prod.

from https://github.com/styled-components/styled-components/issues/1045#issuecomment-319359844 it seems like it should be prefixed but doesn't seem to be happening.

related about a bug in an older version but we're on a version newer than that:
https://github.com/thysultan/stylis/issues/75#issuecomment-346366676